% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe_index.R
\name{Phe_index}
\alias{Phe_index}
\title{phe_index}
\usage{
Phe_index(dat, pos_spe, neg_spe)
}
\arguments{
\item{dat, }{data frame for the metagenome data, row is sample and col is viarable}

\item{pos_spe, }{a vector of features which are positive association with phenotype}

\item{neg_spe, }{a vector of features which are negative association with phenotype}
}
\value{
data.frame including index
}
\description{
This function is provided for compute the combined index for a spefic phenotype
}
\examples{
data(MH)
data(MN)
data(sp_data)
sp_data_t <- t(sp_data)
sp_data_n <- sp_data_t/rowSums(sp_data_t)
GMHI <- Phe_index(sp_data_n, MH, MN)
}
\references{
Gupta VK, Kim M, Bakshi U, et al. A predictive index for health status using species-level gut microbiome profiling. Nat Commun. 2020;11(1):4635.
}
