% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mdiversity.R
\name{Mdiversity}
\alias{Mdiversity}
\title{Integrated ecological diversity indexes}
\usage{
Mdiversity(x, index = "shannon", margin = 1, ...)
}
\arguments{
\item{x}{Community data}

\item{index}{diversity index, one of "richness", "shannon", "simpson",
"invsimpson","uniquess_jaccard", "uniquess_bray", "uniquess_aitchison",
"uniquess_kendel"; "all" indicates caluated all these index}

\item{margin}{row or col}

\item{...}{lev parameter for uniquess function, indicates threshold to compute the uniquess}
}
\value{
data.frame
}
\description{
Integrated ecological diversity indexes
}
